requirements.txt
setuptools==59.6.0
pip==20.3.4
asgiref==3.3.4
asn1crypto==0.24.0
awscli==1.19.98
awsebcli==3.20.0
bcrypt==3.2.0
beautifulsoup4==4.7.1
blessed==1.15.0
boto==2.49.0
boto3==1.17.98
boto3-stubs==1.17.98
botocore==1.20.98
cached-property==1.5.1
cement==2.8.2
certifi==2018.8.24
cffi==1.14.5
chardet==3.0.4
cloudipsp==1.0.2
colorama==0.3.9
dj-database-url==0.5.0
Django==1.11
django-autoslug==1.9.3
django-autoslug-iplweb==1.9.4.dev0
cryptography==3.4.7
django-cognito-jwt==0.0.3
django-cors-headers==2.4.0
django-elasticsearch-dsl==0.5.1
django-elastipymemcache==1.2.1
django-health-check==3.12.1
django-hvad==1.8.0
django-pylibmc==0.6.1
django-ses==1.0.3
django-smuggler==0.9.1
django-storages==1.7.1
django-stubs==1.5.0
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
docker==3.7.0
docker-compose==1.25.5
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ecdsa==0.13
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
future==0.16.0
html2text==2020.1.16
idna==2.7
ipaddress==1.0.22
jmespath==0.9.4
jsonschema==2.6.0
mypy==0.770
mypy-extensions==0.4.3
paramiko==2.7.2
pathspec==0.5.9
Pillow==7.2.0
pipreqs==0.4.9
progressbar2==3.47.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pycparser==2.18
pycryptodomex==3.4.7
PyJWT==1.6.4
pylibmc==1.6.1
pymemcache==3.2.0
PyNaCl==1.4.0
python-dateutil==2.8.0
python-decouple==3.1
python-jose==3.0.0
python-utils==2.3.0
pytz==2018.5
PyYAML==5.3.1
requests==2.25.1
requests-aws4auth==0.9
rsa==3.4.2
s3transfer==0.4.2
semantic-version==2.8.5
six==1.11.0
sorl-thumbnail==12.5.0
soupsieve==1.9.1
sqlparse==0.4.1
stream-django==1.4.0
stream-python==3.0.1
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
Unidecode==1.0.22
Unipath==1.1
urllib3==1.25.4
wcwidth==0.1.7
websocket-client==0.55.0
yarg==0.1.9

